+++ src/server/index.js
@@ -1,0 +1,5 @@
+import 'babel-polyfill';
+import start from './start';
+
+// Server API.
+export default { start };